Make Mpg321 a module, add Mpg321::Client class.
