+++ lib/mpg321/version.rb
@@ -1,3 +1,3 @@
-class Mpg321
+module Mpg321
   VERSION = '0.4.0'
 end
